requirements.txt
fastapi
uvicorn
torch
torchvision
Pillow
requests
pydantic
numpy
